Convert Name#references to a service object
