+++ app/models/name.rb
@@ -90,7 +90,7 @@ class Name < ApplicationRecord
   end
 
   def references
-    references_in_fields.concat(references_in_taxt)
+    Names::WhatLinksHere.new(self).call
   end
 
   private
@@ -101,50 +101,6 @@ class Name < ApplicationRecord
     def set_taxon_caches
       Taxon.where(name: self).update_all(name_cache: name)
       Taxon.where(name: self).update_all(name_html_cache: name_html)
-    end
-
-    def references_in_fields
-      references_to_taxon_name
-        .concat(references_to_taxon_type_name)
-        .concat(references_to_protonym_name)
-    end
-
-    def references_to_taxon_name
-      Taxon.where(name: self).map do |taxon|
-        table_ref 'taxa', :name_id, taxon.id
-      end
-    end
-
-    def references_to_taxon_type_name
-      Taxon.where(type_name: self).map do |taxon|
-        table_ref 'taxa', :type_name_id, taxon.id
-      end
-    end
-
-    def references_to_protonym_name
-      Protonym.where(name: self).map do |protonym|
-        table_ref 'protonyms', :name_id, protonym.id
-      end
-    end
-
-    def references_in_taxt
-      references = []
-      Taxt::TAXT_FIELDS.each do |klass, fields|
-        table = klass.arel_table
-        fields.each do |field|
-          klass.where(table[field].matches("%{nam #{id}}%")).each do |record|
-            next unless record[field]
-            if record[field] =~ /{nam #{id}}/
-              references << table_ref(klass.table_name, field, record.id)
-            end
-          end
-        end
-      end
-      references
-    end
-
-    def table_ref table, field, id
-      { table: table, field: field, id: id }
     end
 
     def change name_string
